refactor(parity): split unported-files.ts into a per-package directory
Mechanical move only. `scripts/parity/unported-files.ts` becomes
`scripts/parity/unported-files/`: one module per `package` value, plus
`unscoped.ts` for the 106 entries that carry no `package` (they match
across every package, which is exactly why they cannot be filed under a
package name), `types.ts` as the single home for the schema comment and
the `UnportedFile` type, and `index.ts` merging the shards and owning
the three `is*Unported` predicates verbatim.

No entry gains or loses a `package` field, no reason string is reworded,
no entry is edited. `unported-files/baseline.json` snapshots the
pre-split array and a new test asserts the merged register reproduces it
entry-for-entry.

Story: 0097-parity-output-sharding/unported-files-split-per-package
