+++ scripts/parity/unported-files/unscoped.ts
@@ -1,56 +1,13 @@
-// Ruby files intentionally excluded from parity:api / parity:test.
+// Entries with no `package` field: they match across ALL packages.
+// This file is where an entry lives precisely BECAUSE it carries no
+// package scope — do not add `package:` to an entry to move it out of
+// here, that narrows what parity:api / parity:test exclude.
 //
-// Two kinds of exclusions:
-//   - pre-1.0 scope: features we haven't committed to porting yet
-//     (migration compatibility shims, legacy adapters, etc.)
-//   - not-applicable: Ruby-only concerns that don't map to JS
-//     (thread-pool plumbing, Marshal/Psych/MessagePack formats, etc.)
-//
-// Each entry must set at least one of:
-//   `pattern`  — substring match against the Ruby SOURCE file path
-//                (from extract-ruby-api.rb, e.g. "promise.rb").
-//                Consumed by isSourceUnported() → parity:api.
-//                A leading "/" anchors it to a path boundary, exactly as in
-//                `testFile` below: "/version.rb" matches the top-level
-//                `version.rb` and `<dir>/version.rb`, but NOT `gem_version.rb`.
-//                Omit for test-only entries where the source IS being ported.
-//   `testFile` — substring match against the Ruby TEST file path
-//                (from extract-ruby-tests.rb, e.g. "message_pack_test.rb").
-//                Consumed by isTestFileUnported() → parity:test.
-//                Omit when there is no corresponding Rails test file.
-//   `package`  — (optional) scopes a `pattern` (source-path) or whole-file
-//                `testFile` (test-path) match to one package. Required when the
-//                same basename exists in more than one package — e.g.
-//                `core_ext/name_error.rb` lives in both activesupport and
-//                did_you_mean, and `railtie_test.rb` at the test-root of both
-//                globalid and activemodel — where the exclusion applies to only
-//                one. Unscoped entries match across all packages.
-//
-// Most entries set both (source and test excluded together).
-// Test-only entries (GVL, Rake, dbconsole, Ruby serialization) set only
-// `testFile` because their TS source counterparts either don't exist or
-// are being actively ported.
+// See ./types.ts for the entry schema.
 
-export type UnportedFile = { reason: string } &
-  // `package` scopes a source-path (`pattern`) or whole-file test-path
-  // (`testFile` without `tests`) match to one package. Per-test entries
-  // (`tests:`) match on the test description, so scoping is pointless there.
-  (| { pattern: string; testFile?: string; tests?: never; package?: string }
-    | { pattern?: string; testFile: string; tests?: never; package?: string }
-    // `className` narrows a per-test exclusion to one Ruby *Test class within
-    // the file, so a GVL-only subclass can be dropped while a portable sibling
-    // sharing a test name stays counted. Per-test entries never touch parity:api.
-    | { pattern?: never; testFile: string; className?: string; tests: string[]; package?: never }
-  );
+import type { UnportedFile } from "./types.js";
 
-export const UNPORTED_FILES: UnportedFile[] = [
-  {
-    pattern: "i18n_railtie.rb",
-    package: "activesupport",
-    reason:
-      "Rails boot lifecycle hook that wires I18n into the Railtie/Application init sequence. " +
-      "No JS equivalent — I18n is configured directly in user code.",
-  },
+export const UNSCOPED_UNPORTED_FILES: UnportedFile[] = [
   {
     pattern: "migration/compatibility", // test excluded by extract-ruby-tests.rb SKIP_PATTERNS (/\/migration\//)
     reason: "Pre-1.0: legacy Rails version migration compatibility shims.",
@@ -665,23 +622,6 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "protocol; callers read `.app` / `.modelName` / `.modelId` / `.params` " +
       "directly off the GID instance.",
   },
-  // --- globalid: Rails::Railtie wiring ---
-  {
-    testFile: "/railtie_test.rb",
-    package: "globalid",
-    reason:
-      "Exercises the `global_id` Railtie initializer through a full " +
-      "`Rails::Application` boot (`@app.initialize!`) under " +
-      "`ActiveSupport::Testing::Isolation` — GlobalID.app defaulting, " +
-      "`config.global_id.app`/`expires_in` injection, and verifier key " +
-      "derivation from `app.key_generator`. Unlike activemodel/trailties, " +
-      "globalid has not yet ported its railtie to a `Trailtie` " +
-      "(activesupport's `BaseRailtie`); its wiring is a `wire.ts` side-effect " +
-      "plus explicit `setApp`/verifier setters, and there is no " +
-      "`Rails::Application` boot harness for these tests to drive. Porting " +
-      "globalid's railtie to a `Trailtie` would let this file re-enter " +
-      "accounting — tracked as a follow-up story.",
-  },
   // --- globalid: module-based `only:` filters (Ruby modules have no TS analogue) ---
   {
     testFile: "global_locator_test.rb",
@@ -753,101 +693,6 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "imports hits the TS private-field nominal-typing trap; the " +
       "`SignedGlobalID#equals` contract is symmetric only within its own " +
       "class. Same-class equality is covered by `value equality`.",
-  },
-  // --- did-you-mean: Ruby-only checkers ---
-  // The DidYouMean port is scoped to the algorithms Rails consumes
-  // (SpellChecker + Jaro/JaroWinkler/Levenshtein). The remaining files
-  // patch Ruby's exception hierarchy (NameError#corrections via
-  // core_ext/name_error.rb + formatter.rb), or implement checkers that
-  // suggest names for Ruby-only failure modes — NoMethodError method
-  // names, NameError class/variable names, KeyError keys,
-  // NoMatchingPatternKeyError keys, $LOAD_PATH require targets. None of
-  // these map onto JS errors.
-  {
-    package: "did-you-mean",
-    pattern: "core_ext/name_error.rb",
-    testFile: "core_ext/test_name_error_extension.rb",
-    reason:
-      "Patches Ruby's NameError with `corrections`, `original_message`, " +
-      "`detailed_message`, `spell_checker`. JS has no NameError; trails " +
-      "errors carry their own per-subclass `corrections` getter (see " +
-      "ActionNotFound, ParameterMissing, AssociationNotFoundError, etc.).",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "formatter.rb",
-    reason:
-      "Formats `Did you mean? …` suffix for Ruby's Exception#detailed_message " +
-      "integration. JS error stringification is per-error, not via a stdlib hook.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/key_error_checker.rb",
-    testFile: "spell_checking/test_key_name_check.rb",
-    reason: "Suggests Hash/ENV keys on Ruby KeyError. JS Map/object access doesn't raise.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/method_name_checker.rb",
-    testFile: "spell_checking/test_method_name_check.rb",
-    reason:
-      "Suggests method names on Ruby NoMethodError using receiver.methods. " +
-      "JS has no NoMethodError; undefined property access returns undefined.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/name_error_checkers.rb",
-    reason: "Dispatcher for class/variable name checkers; both checkers are Ruby-only (see below).",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/name_error_checkers/class_name_checker.rb",
-    testFile: "spell_checking/test_class_name_check.rb",
-    reason:
-      "Suggests constant/class names by walking Module#constants + ancestor scopes. " +
-      "Ruby-only — JS has no equivalent constant introspection.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/name_error_checkers/variable_name_checker.rb",
-    testFile: "spell_checking/test_variable_name_check.rb",
-    reason:
-      "Suggests local/instance/class variable names from Binding#local_variables, " +
-      "Object#instance_variables, etc. Ruby-only introspection surface.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/null_checker.rb",
-    testFile: "spell_checking/test_uncorrectable_name_check.rb",
-    reason:
-      "Null-object fallback in Ruby's checker registry. Not needed in our " +
-      "error-subclass approach.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/pattern_key_name_checker.rb",
-    testFile: "spell_checking/test_pattern_key_name_check.rb",
-    reason:
-      "Suggests keys on Ruby NoMatchingPatternKeyError (one-liner pattern matching). " +
-      "No JS equivalent.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "spell_checkers/require_path_checker.rb",
-    testFile: "spell_checking/test_require_path_check.rb",
-    reason:
-      "Suggests $LOAD_PATH targets on Ruby LoadError. JS module resolution is " +
-      "engine-handled and doesn't surface this kind of typo suggestion.",
-  },
-  {
-    package: "did-you-mean",
-    pattern: "tree_spell_checker.rb",
-    testFile: "test_tree_spell_checker.rb",
-    reason:
-      "Path-segment spell checker used by Rails::TestUnit::InvalidTestError to " +
-      "suggest test file paths. Pre-1.0 scope: trails doesn't yet port " +
-      "`bin/rails test` runner wiring, and the algorithm has no other consumer " +
-      "in the call sites we target.",
   },
   {
     testFile: "tree_spell/test_change_word.rb",
@@ -1132,139 +977,6 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "See packages/activerecord/src/database-configurations/resolver.test.ts.",
   },
   {
-    pattern: "tools.rb",
-    package: "activerecord-test-support",
-    reason:
-      "Rails' rake/minitest test-runner plumbing: prepends adapter globbing onto " +
-      "Rails::TestUnit::Runner and invokes it with ARGV. vitest fills that role in " +
-      "trails — no port intended.",
-  },
-  // --- i18n: optional backend mixins composed on top of Simple ---
-  {
-    pattern: "backend/cache.rb",
-    testFile: "backend/cache_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — caches lookups through ActiveSupport::Cache.",
-  },
-  {
-    pattern: "backend/cache_file.rb",
-    testFile: "backend/cache_file_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — caches parsed translation files on disk.",
-  },
-  {
-    pattern: "backend/cascade.rb",
-    testFile: "backend/cascade_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — cascading key lookup (`foo.bar.baz` → `foo.baz`).",
-  },
-  {
-    pattern: "backend/interpolation_compiler.rb",
-    testFile: "backend/interpolation_compiler_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — compiles interpolations into Ruby " +
-      "procs via `eval`/`instance_eval`, which has no trails counterpart.",
-  },
-  {
-    pattern: "backend/lazy_loadable.rb",
-    testFile: "backend/lazy_loadable_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — defers translation-file loading until first lookup.",
-  },
-  {
-    pattern: "backend/memoize.rb",
-    testFile: "backend/memoize_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — memoizes lookups per locale.",
-  },
-  {
-    pattern: "backend/metadata.rb",
-    testFile: "backend/metadata_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — attaches lookup metadata onto the " +
-      "translated String's singleton class. JS strings take no per-instance state.",
-  },
-  {
-    pattern: "backend/pluralization.rb",
-    testFile: "backend/pluralization_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — pluralization via per-locale rule procs in the data.",
-  },
-  {
-    testFile: "backend/pluralization_fallback_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — pluralization via per-locale rule procs in the data. " +
-      "Suite for the deferred `backend/pluralization.rb` composed with Fallbacks.",
-  },
-  {
-    testFile: "backend/pluralization_scope_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — pluralization via per-locale rule procs in the data. " +
-      "Suite for the deferred `backend/pluralization.rb` looked up under a scope.",
-  },
-  // --- i18n: api/ suites that install a deferred backend ---
-  {
-    testFile: "api/all_features_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixins — a Chain of every deferred mixin " +
-      "(Cascade, Memoize, Metadata, Pluralization, Fallbacks) at once.",
-  },
-  {
-    testFile: "api/cascade_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — cascading key lookup (`foo.bar.baz` → `foo.baz`).",
-  },
-  {
-    testFile: "api/lazy_loadable_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — defers translation-file loading until first lookup.",
-  },
-  {
-    testFile: "api/memoize_test.rb",
-    package: "i18n",
-    reason: "Pre-1.0: optional backend mixin — memoizes lookups per locale.",
-  },
-  {
-    testFile: "api/pluralization_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: optional backend mixin — pluralization via per-locale rule procs in the data.",
-  },
-  // --- i18n: gem surface trails has no counterpart for ---
-  {
-    // Deliberately broad: covers `gettext.rb`, `gettext/*` and the
-    // `backend/gettext.rb` mixin in one entry.
-    pattern: "gettext",
-    testFile: "gettext/",
-    package: "i18n",
-    reason:
-      "Pre-1.0: gettext .po support — the catalogue parser, the `_`/`n_`/`s_` " +
-      "helpers, and the backend mixin that loads .po files. A Ruby toolchain " +
-      "concern with no trails consumer; Rails' own I18n usage never touches it.",
-  },
-  {
-    testFile: "gettext_plural_keys_test.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: gettext .po support — exercises `I18n::Gettext.plural_keys`. " +
-      "Sits at test/i18n/, outside the `gettext/` prefix above.",
-  },
-  {
-    pattern: "middleware.rb",
-    testFile: "i18n/middleware_test.rb",
-    package: "i18n",
-    reason:
-      "Rack middleware that resets `I18n.locale` after each request. trails " +
-      "has no Rack request lifecycle wiring for i18n yet; resetting is done " +
-      "explicitly by callers.",
-  },
-  {
     testFile: "i18n_test.rb",
     tests: ["exposes its VERSION constant"],
     reason:
@@ -1290,39 +1002,6 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "engine raising `ArgumentError: invalid byte sequence` (transliterator.rb:96). A JS " +
       "string is UTF-16 code units with no invalid-byte state to reach, so no input makes " +
       "the ported `transliterate` raise.",
-  },
-  {
-    pattern: "tests/",
-    package: "i18n",
-    reason:
-      "`lib/i18n/tests/*` are minitest mixins the gem ships so third-party " +
-      "backends can run its conformance suite. Test-support scaffolding, not " +
-      "library surface — trails' backend tests are vitest files instead.",
-  },
-  {
-    pattern: "tests.rb",
-    package: "i18n",
-    reason:
-      "Pre-1.0: the `I18n::Tests` namespace's autoload shim for the `tests/` " +
-      "conformance mixins excluded above — a file that sits beside the " +
-      "directory the pattern above reaches.",
-  },
-  {
-    testFile: "test_date_ractor.rb",
-    package: "date",
-    reason:
-      "Ractor is Ruby's actor-based parallelism primitive — the file exercises " +
-      "`Ractor.make_shareable`/`Ractor.new` round-tripping of a Date across " +
-      "isolated interpreter states. JS is single-threaded and has no analogue, " +
-      "same reason `promise.rb` is excluded above.",
-  },
-  {
-    testFile: "test_date_marshal.rb",
-    package: "date",
-    reason:
-      "Ruby's Marshal binary object format — `Marshal.dump`/`Marshal.load` of a " +
-      "Date, including the frozen-string and instance-variable round trip. " +
-      "Not applicable: JS has no Marshal, and the wire format is Ruby-only.",
   },
   {
     testFile: "test_date_new.rb",
@@ -1351,76 +1030,6 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "rest of the file stays counted.",
   },
   {
-    pattern: "testing/error_reporter_assertions.rb",
-    package: "activesupport",
-    reason:
-      "Minitest assertion module (`assert_error_reported`) for Rails' own test suite; " +
-      "trails has no port, and its `subscribe`/`record`/`report` names otherwise cluster " +
-      "onto the unrelated `error-reporter.ts`.",
-  },
-  // The activesupport `core_ext/*` tail with no trails counterpart (RFC 0072).
-  {
-    pattern: "core_ext/module/attribute_accessors_per_thread.rb",
-    package: "activesupport",
-    reason: "Stores the value per `Thread.current`; JS has no thread-local storage.",
-  },
-  {
-    pattern: "core_ext/module/remove_method.rb",
-    package: "activesupport",
-    reason: "Method-table surgery to silence redefinition warnings; JS reassignment is silent.",
-  },
-  {
-    pattern: "starts_ends_with.rb",
-    package: "activesupport",
-    reason: "Aliases `start_with?`/`end_with?` for Symbol and String; both are native JS.",
-  },
-  {
-    pattern: "core_ext/string/multibyte.rb",
-    package: "activesupport",
-    reason:
-      "`mb_chars` returns a Multibyte::Chars proxy and `is_utf8?` reports a Ruby Encoding; " +
-      "JS strings carry no encoding tag and are already Unicode.",
-  },
-  {
-    pattern: "core_ext/kernel/singleton_class.rb",
-    package: "activesupport",
-    reason:
-      "Ruby metaprogramming with no JS equivalent; trails assigns to the class object directly.",
-  },
-  {
-    pattern: "core_ext/pathname/existence.rb",
-    package: "activesupport",
-    reason:
-      "Ruby's Pathname is not ported (trails paths are strings) and the check is async in JS.",
-  },
-  {
-    pattern: "core_ext/regexp.rb",
-    package: "activesupport",
-    reason: "Reads the `//m` bit out of Ruby's options bitmask; JS `RegExp` exposes `.multiline`.",
-  },
-  {
-    pattern: "core_ext/integer/multiple.rb",
-    package: "activesupport",
-    reason: "Pre-1.0: `Integer#multiple_of?` is unported and has no trails caller.",
-  },
-  {
-    pattern: "core_ext/module/deprecation.rb",
-    package: "activesupport",
-    reason:
-      "Pre-1.0: class-body sugar over `Deprecation#deprecate_methods`; trails calls the " +
-      "deprecator directly (`deprecation.ts#deprecateMethod`).",
-  },
-  {
-    pattern: "core_ext/object/with_options.rb",
-    package: "activesupport",
-    reason: "Pre-1.0: needs ActiveSupport::OptionMerger, not ported yet (RFC 0093).",
-  },
-  {
-    pattern: "core_ext/string/behavior.rb",
-    package: "activesupport",
-    reason: "Pre-1.0: one arm of `acts_like?`; trails ports the Time/Date arms only.",
-  },
-  {
     pattern: "/version.rb",
     reason:
       "`Module.version` / `VERSION` returns the gem version as a Gem::Version " +
@@ -1431,49 +1040,3 @@ export const UNPORTED_FILES: UnportedFile[] = [
       "(`ActionPack.gem_version`).",
   },
 ];
-
-export function isSourceUnported(file: string, pkg?: string): boolean {
-  return UNPORTED_FILES.some((e) => {
-    if (!e.pattern) return false;
-    // A leading "/" anchors the pattern to a path boundary — same rule as
-    // `isTestFileUnported` below — so a basename entry cannot swallow a longer
-    // basename that ends with it (`version.rb` vs `gem_version.rb`).
-    const matched = e.pattern.startsWith("/")
-      ? file === e.pattern.slice(1) || file.includes(e.pattern)
-      : file.includes(e.pattern);
-    if (!matched) return false;
-    // `package` only exists on the pattern-bearing variant of the union.
-    const scope = "package" in e ? e.package : undefined;
-    return scope === undefined || pkg === undefined || scope === pkg;
-  });
-}
-
-export function isTestFileUnported(testFile: string, pkg?: string): boolean {
-  return UNPORTED_FILES.some((e) => {
-    if (!e.testFile || e.tests) return false;
-    const scope = "package" in e ? e.package : undefined;
-    if (scope !== undefined && pkg !== undefined && scope !== pkg) return false;
-    const p = e.testFile;
-    // A leading "/" anchors the pattern to a path boundary: match the exact
-    // basename (`testFile === p.slice(1)`, for top-level files with no dir
-    // prefix) or a literal substring with the slash (`testFile.includes(p)`,
-    // for files under a subdir).  Without a leading "/", plain substring.
-    if (p.startsWith("/")) return testFile === p.slice(1) || testFile.includes(p);
-    return testFile.includes(p);
-  });
-}
-
-export function isTestCaseUnported(
-  testFile: string,
-  testName: string,
-  className?: string,
-): boolean {
-  return UNPORTED_FILES.some(
-    (e) =>
-      e.testFile &&
-      e.tests &&
-      testFile.includes(e.testFile) &&
-      e.tests.includes(testName) &&
-      (e.className === undefined || e.className === className),
-  );
-}
